Solving 8 unitAmout with 5:2:1 ratio for creating Receipt Voucher via payment create
Solving 8 unitAmout with 5:2:1 ratio for creating Receipt Voucher via payment create

+++ core-services/egov-user/src/main/java/org/egov/user/security/oauth2/custom/authproviders/CustomAuthenticationProvider.java
@@ -173,6 +173,7 @@ public class CustomAuthenticationProvider implements AuthenticationProvider {
 			 */
 			System.out.println("uuid password valid::"+uuid);
 			//restTemplate.delete(uuid);
+			if (captchaForDev)
 			userService.deleteCaptcha(uuid);
 			List<GrantedAuthority> grantedAuths = new ArrayList<>();
 			grantedAuths.add(new SimpleGrantedAuthority("ROLE_" + user.getType()));
